Added unread conversation count display to "Inbox" tab. Issue #23
Created count_unread_conversations() function in sql_common.php that returns the count of all unread conversations for the session user.

A jquery-ajax function was added in messages.php. This calls count_unread_conversations() from jq_ajax.php and the count returned in displayed in a <span> with id="status_view".

A $_POST["status_to_read"] check was added in message_view.php to change the status of the conversation to read on opening an unread conversation.

Added some visual effects in style.css to change the view of unread messages. Text is displayed BOLD with a green background-color. The "unread" visual class in added to any unread conversations in received_messages.php