Auto-remove deleted conversations after allotted time. Issue #13
Due to a restriction on using events or foreign keys on the current website hosting server, a work around has been added using php function and sql queries.
'destroy' is a new conversation status which is set on any conversation that is meant to be removed from the database at a later time. A 'destroy' conversation is no longer visible to a user and is thus considered 'permanently deleted' on the users end.
When a user deletes a conversation, a new date is stored in the Conversation table under two new columns; sender_destroyDate and receiver_destroyDate.
This is not the deletion date but the date the conversation should be 'destroyed'; have its status changed to 'destroy'.
The date is (date deleted + X days) which is currently set to '1 week' or 7 days; (date deleted + 1 week).
Whenever the user logs in succesfully, a php function is called which proceeds to check if there are any conversations for the session user that should be destroyed on the current date. If any are found, their status is set to 'destroy'.
The conversations become invisible to the user as if permanently deleted. The conversations are not removed from the database.


ims_schema.sql:
 - Two new columns were added in the Conversation table; sender_destroyDate and receiver_destroyDate.
 - Changed table engines to MyISAM.
 - A trigger has also been added to replace a foreign key which updates usernames in Conversation and Message table if a user changed their username in User table.

sql_common.php:
 - Added set_destroy_date() function. It sets the date for when a conversation should be destroyed.
 - Added set_destroy_status() function. It sets the status of the conversation to 'destroy' when the date set in sender_destroyDate or receiver_destroyDate is reached.
 - Updated change_conversation_status to a single query and simpler logic.

received_messages.php:
- Added set_destroy_date() function after a successful IF check on change_conversation_status(). This deletes a conversation and sets the date for when to destroy.

deleted_messages.php:
- Added set_destroy_date() function after a successful IF check on change_conversation)status(). This undeletes a conversation and sets the date to destroy the conversation to NULL.

message_view.php:
- Added set_destroy_date() function to set the destroy date to NULL if a new message is sent to a deleted or destroyed conversation by the other user in a conversation.

login.php:
- Added the set_destroy_status() function on successful login check. This checks if any conversations should be set to 'destroy' if their destroy date has reached.