Added new view for 'add item' section. Issue #51
The 'add item' section is now hidden in a sliding drawer at the top of the page.

Included a new field for adding quantity to the new item. Issue #62

A new input field has been added in the 'add item' section for adding item quantity. The function set_base_quantity() is called to set the new value.