Added pagination for items table. Issue #31
The items table is now paginated with 10 items per page.
The pagination row is displayed about the table with page displayed as a separate number block.
Pages can be traversed using the arrow keys or jump to the page that is clicked on.

The new page table data is called through jquery ajax and displayed via javascript.

Currently the 10 items per page limit can not be changed by the user.